+++ source/graph/graph.h
@@ -7,7 +7,7 @@ typedef struct graph Graph;
 
 Graph *graph_create(int V, int E);
 
-void node_insert(Graph * g, int a, int b, float distance);
+void node_insert(Graph * g, int a, int b, double distance);
 
 void graph_print(Graph* graph);
